requirements.txt
beautifulsoup4==4.6.0
bs4==0.0.1
pbkdf2==1.3
pycrypto==2.6.1
pystache==0.5.4
